Fix debug printing function: belle_sip_debug().
The original code tried to map belle_sip_deb(...) (i.e., a variadic
macro) to bctbx_debug(...) (i.e., another variadic macro as the
replacement).

This didn't work for multiple reasons:
  - no code used belle_sip_deb(), but actually belle_sip_debug().
  - the target function has to be called with either a named variadic
    argument (e.g., "nameargs..." -> "nameargs") or, if it's
    unnamed/anonymous via the portable __VA_ARGS__ name/macro.

Rename it to belle_sip_debug and make it a non-variadic substitution to
bctbx_debug, since the usage is 1:1 currently.

If it stops being a 1:1 replacement in the future, it's easy to just
make it variadic again such as via:

  belle_sip_debug(args...) some_func(NULL, args)