Upon discovery of a public IP address, copy it to channels that have none.
This fixes the infamous "register loop" issue that is, e.g., triggered
by networks with multi-homed hosts which also act as SIP servers.

Consider this example:
  - client [10.0.0.2] contacts
    server [10.0.0.1 (N.B.: private), 203.0.113.14 (N.B.: public)]
    by using [203.0.113.14] explicitly.
  - belle-sip channel [NULL, 203.0.113.14] is being created (since the
    client-side public IP address can only be discovered on replies).
  - server responds via [10.0.0.1] instead.
  - belle-sip channel [10.0.0.2, 10.0.0.1] is being created (client-side
    public IP address comes from SIP packet information).
  - client notices that the pair (NULL, [10.0.0.2]) doesn't match,
    assumes that REGISTER failed.
  - new REGISTER round starts.

Applying the newly learned public IP address to other channels that have
none set should, in theory, only ever change one channel - the
corresponding sending socket.

Practically, though, depending on factors such as if belle-sip can be
used in a multi-threaded fashion or if multiple connections can be
opened (almost) at the same time and network timing effects, it's not
unthinkable of situations like these:
  - client sends REGISTER to server1 and waits => channel1
  - client sends REGISTER to server2 and waits => channel2
  - server2 answers via different IP address => channel3 => public IP
    address passed to channel1 (wrongly) and channel2 (correctly)
  - server1 answers via different IP address => channel4
  - public IP address pair of channel1 and channel4 won't match =>
    infinite REGISTER sequence starts

It's unclear to me if this race can happen, though. Fixing it would be
very difficult, too, since we'd have to synchronize requests and
replies, which currently does not seem to be a concept in belle-sip, or
to match sending and receiving channels, which is not trivially possible
and might even be impossible with more work involved.

See https://github.com/BelledonneCommunications/linphone-sdk/issues/17