Skip CRLF at the start of a NOTIFY body
https://www.w3.org/Protocols/rfc1341/7_2_Multipart.html

"The requirement that the encapsulation boundary begins with a CRLF
implies that the body of a multipart entity must itself begin with a
CRLF before the first encapsulation line -- that is, if the "preamble"
area is not used, the entity headers must be followed by TWO CRLFs. This
is indeed how such entities should be composed. A tolerant mail reading
program, however, may interpret a body of type multipart that begins
with an encapsulation line NOT initiated by a CRLF as also being an
encapsulation boundary, but a compliant mail sending program must not
generate such entities."