Add Multipath TCP (MPTCP) support
Multipath TCP or MPTCP is an extension to the standard TCP and is described in RFC 8684.
It allows a device to make use of multiple interfaces at once to send and receive TCP packets over a single MPTCP connection.
MPTCP can aggregate the bandwidth of multiple interfaces or prefer the one with the lowest latency.
It also allows a fail-over if one path is down, and the traffic is seamlessly reinjected on other paths.

For more details about MPTCP, refer to the official website https://mptcp.dev

Typical uses cases for this library would be to leverage the fail-over capacity of MPTCP to
enhance the user experience on mobile devices (android/iOS).

Changes to this repository consist of replacing the protocol field value, previously IPPROTO_TCP,
by IPPROTO_MPTCP (if this symbol is available) when creating sockets.