Allow to disable RFC5626 PING PONG sanity check
By default belle-sip waits for the first PONG response to ensure that
the server supports the RFC5626. If is not received within the timeout,
it will be assumed thazt RFC5626 is not supported by the server.

Now the user can choose to enable or disable such a sanity check as it
may lead to false negatives if the server is quite slow in sending the
PONG or the timeout is not big enough